Added basic Keybind setting and Module setting tab.
The new Binding system saves the Binds to the new Settings file. This means that any changes made to bindings will persist across game sessions, similar to other module settings. There is now a Module Settings Tab instead of a dropdown in the Module Component. This is likely to change in the future but it's a good placeholder.