Many UI QoL fixes. Added new Color Picker component along with new HUD Icons.
Included are many QoL fixes that make it easier to navigate the UI.

There is a new Color Picker component that allows you to visually change the colour of your UI, rather than using the Hue color picker. There will be improvements made but this is a good start.

Changes were made to all components to work properly with stack panels.