More UI refactoring.
Added in a ItemsComponent, which allows dynamic creation of UI elements based off of a List of items. Can be used to show repeating components that represent some form of data (ie alts, macros, friends, etc...)
Added in more components on the Macro screen.
Component's parent is now assumed when it is added to the parent, reducing the potential for the develop to mess up UI.
Other small tweaks to make order of operations more concise.